# Optimize KnownUnreliableOrOffline SQL query