--- a/pkg/overlay/benchmark_test.go
+++ b/pkg/overlay/benchmark_test.go
@@ -9,12 +9,9 @@ import (
 	"testing"
 	"time"
 
-	"github.com/stretchr/testify/assert"
+	"github.com/golang/protobuf/ptypes"
 	"github.com/stretchr/testify/require"
-	"github.com/zeebo/errs"
 
-	"storj.io/storj/internal/testcontext"
-	"storj.io/storj/internal/testplanet"
 	"storj.io/storj/pkg/overlay"
 	"storj.io/storj/pkg/pb"
 	"storj.io/storj/pkg/storj"
@@ -22,10 +19,10 @@ import (
 	"storj.io/storj/satellite/satellitedb/satellitedbtest"
 )
 
-func BenchmarkOffline(b *testing.B) {
+func BenchmarkOverlay(b *testing.B) {
 	satellitedbtest.Bench(b, func(b *testing.B, db satellite.DB) {
 		const (
-			TotalNodeCount = 200
+			TotalNodeCount = 211
 			OnlineCount    = 90
 			OfflineCount   = 10
 		)
@@ -33,19 +30,20 @@ func BenchmarkOffline(b *testing.B) {
 		overlaydb := db.OverlayCache()
 		ctx := context.Background()
 
+		var all []storj.NodeID
 		var check []storj.NodeID
 		for i := 0; i < TotalNodeCount; i++ {
 			var id storj.NodeID
 			_, _ = rand.Read(id[:]) // math/rand never returns error
-
-			err := overlaydb.UpdateAddress(ctx, &pb.Node{
-				Id: id,
-			})
-			require.NoError(b, err)
-
+			all = append(all, id)
 			if i < OnlineCount {
 				check = append(check, id)
 			}
+		}
+
+		for _, id := range all {
+			err := overlaydb.UpdateAddress(ctx, &pb.Node{Id: id})
+			require.NoError(b, err)
 		}
 
 		// create random offline node ids to check
@@ -55,20 +53,72 @@ func BenchmarkOffline(b *testing.B) {
 			check = append(check, id)
 		}
 
-		criteria := &overlay.NodeCriteria{
-			AuditCount:         0,
-			AuditSuccessRatio:  0.5,
-			OnlineWindow:       1000 * time.Hour,
-			UptimeCount:        0,
-			UptimeSuccessRatio: 0.5,
-		}
+		b.Run("KnownUnreliableOrOffline", func(b *testing.B) {
+			criteria := &overlay.NodeCriteria{
+				AuditCount:         0,
+				AuditSuccessRatio:  0.5,
+				OnlineWindow:       1000 * time.Hour,
+				UptimeCount:        0,
+				UptimeSuccessRatio: 0.5,
+			}
+			for i := 0; i < b.N; i++ {
+				badNodes, err := overlaydb.KnownUnreliableOrOffline(ctx, criteria, check)
+				require.NoError(b, err)
+				require.Len(b, badNodes, OfflineCount)
+			}
+		})
 
-		b.ResetTimer()
-		defer b.StopTimer()
-		for i := 0; i < b.N; i++ {
-			badNodes, err := overlaydb.KnownUnreliableOrOffline(ctx, criteria, check)
-			require.NoError(b, err)
-			require.Len(b, badNodes, OfflineCount)
-		}
+		b.Run("UpdateAddress", func(b *testing.B) {
+			for i := 0; i < b.N; i++ {
+				id := all[i%len(all)]
+				err := overlaydb.UpdateAddress(ctx, &pb.Node{Id: id})
+				require.NoError(b, err)
+			}
+		})
+
+		b.Run("UpdateStats", func(b *testing.B) {
+			for i := 0; i < b.N; i++ {
+				id := all[i%len(all)]
+				_, err := overlaydb.UpdateStats(ctx, &overlay.UpdateRequest{
+					NodeID:       id,
+					AuditSuccess: i&1 == 0,
+					IsUp:         i&2 == 0,
+				})
+				require.NoError(b, err)
+			}
+		})
+
+		b.Run("UpdateNodeInfo", func(b *testing.B) {
+			now := ptypes.TimestampNow()
+			for i := 0; i < b.N; i++ {
+				id := all[i%len(all)]
+				_, err := overlaydb.UpdateNodeInfo(ctx, id, &pb.InfoResponse{
+					Type: pb.NodeType_STORAGE,
+					Operator: &pb.NodeOperator{
+						Wallet: "0x0123456789012345678901234567890123456789",
+						Email:  "a@example.com",
+					},
+					Capacity: &pb.NodeCapacity{
+						FreeBandwidth: 1000,
+						FreeDisk:      1000,
+					},
+					Version: &pb.NodeVersion{
+						Version:    "1.0.0",
+						CommitHash: "0",
+						Timestamp:  now,
+						Release:    false,
+					},
+				})
+				require.NoError(b, err)
+			}
+		})
+
+		b.Run("UpdateUptime", func(b *testing.B) {
+			for i := 0; i < b.N; i++ {
+				id := all[i%len(all)]
+				_, err := overlaydb.UpdateUptime(ctx, id, i&1 == 0)
+				require.NoError(b, err)
+			}
+		})
 	})
 }