--- a/pkg/storj/node.go
+++ b/pkg/storj/node.go
@@ -7,6 +7,7 @@ import (
 	"crypto/sha256"
 	"crypto/x509/pkix"
 	"database/sql/driver"
+	"encoding/hex"
 	"math/bits"
 
 	"github.com/btcsuite/btcutil/base58"
@@ -178,6 +179,19 @@ func (id *NodeID) Unmarshal(data []byte) error {
 	return err
 }
 
+func (id NodeID) versionByte() byte {
+	return id[NodeIDSize-1]
+}
+
+// unversioned returns the node ID with the version byte replaced with `0`.
+// NB: Legacy node IDs (i.e. pre-identity-versions) with a difficulty less
+// than `8` are unsupported.
+func (id NodeID) unversioned() NodeID {
+	unversionedID := NodeID{}
+	copy(unversionedID[:], id[:NodeIDSize-1])
+	return unversionedID
+}
+
 // Size returns the length of a node ID (implements gogo's custom type interface)
 func (id *NodeID) Size() int {
 	return len(id)
@@ -188,7 +202,7 @@ func (id NodeID) MarshalJSON() ([]byte, error) {
 	return []byte(`"` + id.String() + `"`), nil
 }
 
-// Value set a NodeID to a database field
+// Value converts a NodeID to a database field
 func (id NodeID) Value() (driver.Value, error) {
 	return id.Bytes(), nil
 }
@@ -231,15 +245,37 @@ func (n NodeIDList) Swap(i, j int) { n[i], n[j] = n[j], n[i] }
 // Less implements sort.Interface.Less()
 func (n NodeIDList) Less(i, j int) bool { return n[i].Less(n[j]) }
 
-func (id NodeID) versionByte() byte {
-	return id[NodeIDSize-1]
-}
+// Value converts a NodeIDList to a database field
+func (n NodeIDList) Value() (driver.Value, error) {
+	if n == nil {
+		return nil, nil
+	}
+	if len(n) == 0 {
+		return []byte("{}"), nil
+	}
 
-// unversioned returns the node ID with the version byte replaced with `0`.
-// NB: Legacy node IDs (i.e. pre-identity-versions) with a difficulty less
-// than `8` are unsupported.
-func (id NodeID) unversioned() NodeID {
-	unversionedID := NodeID{}
-	copy(unversionedID[:], id[:NodeIDSize-1])
-	return unversionedID
+	var wp, x int
+	out := make([]byte, 2 + len(n) * (6 + NodeIDSize*2) - 1)
+	
+	x = copy(out[wp:], []byte(`{"\\x`))
+	wp += x
+
+	for i := range n {
+		x = hex.Encode(out[wp:], n[i].Bytes())
+		wp += x
+		if i +1 < len(n) {
+			x = copy(out[wp:], []byte(`","\\x`))
+			wp += x
+		}
+	}
+
+	x = copy(out[wp:], `"}`)
+	wp += x
+
+	if wp != len(out) {
+		println(wp, len(out))
+		panic("unreachable")
+	}
+
+	return out, nil
 }