--- a/satellite/satellitedb/utils.go
+++ b/satellite/satellitedb/utils.go
@@ -4,8 +4,13 @@
 package satellitedb
 
 import (
+	"database/sql/driver"
+	"encoding/hex"
+
 	"github.com/skyrings/skyring-common/tools/uuid"
 	"github.com/zeebo/errs"
+
+	"storj.io/storj/pkg/storj"
 )
 
 // bytesToUUID is used to convert []byte to UUID
@@ -18,4 +23,41 @@ func bytesToUUID(data []byte) (uuid.UUID, error) {
 	}
 
 	return id, nil
+}
+
+type nodeIDsArray storj.NodeIDList
+
+// Value converts a NodeIDList to a postgres array
+func (nodes nodeIDsArray) Value() (driver.Value, error) {
+	if nodes == nil {
+		return nil, nil
+	}
+	if len(nodes) == 0 {
+		return []byte("{}"), nil
+	}
+
+	var wp, x int
+	out := make([]byte, 2+len(nodes)*(6+storj.NodeIDSize*2)-1)
+
+	x = copy(out[wp:], []byte(`{"\\x`))
+	wp += x
+
+	for i := range nodes {
+		x = hex.Encode(out[wp:], nodes[i].Bytes())
+		wp += x
+		if i+1 < len(nodes) {
+			x = copy(out[wp:], []byte(`","\\x`))
+			wp += x
+		}
+	}
+
+	x = copy(out[wp:], `"}`)
+	wp += x
+
+	if wp != len(out) {
+		println(wp, len(out))
+		panic("unreachable")
+	}
+
+	return out, nil
 }