--- a/satellite/satellitedb/utils_test.go
+++ b/satellite/satellitedb/utils_test.go
@@ -11,6 +11,7 @@ import (
 	"github.com/skyrings/skyring-common/tools/uuid"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+
 	"storj.io/storj/pkg/storj"
 )