KnownUnreliableOrOffline is a bottleneck in repair checker.

Use `ANY` check instead of `IN` for Postgres. This results in with the attached benchmark:
```
using id in ($1, ...)     1176302ns
using id = any($1)         763621ns with pq.ByteaArray(nodeIds.Bytes())
using id = any($1)         740142ns
```

_Measurements on a noisy machine._

This doesn't of course solve our problems because 0.7ms for that query is still quite terrible.

Thanks for submitting a PR!

## Code Review Checklist (to be filled out by reviewer)
 - [ ] Does the PR describe what changes are being made?
 - [ ] Does the PR describe why the changes are being made?
 - [ ] Does the code follow [our style guide](https://github.com/storj/docs/blob/master/code/Style.md)?
 - [ ] Does the code follow [our testing guide](https://github.com/storj/docs/blob/master/code/Testing.md)?
 - [ ] Could the PR be broken into smaller PRs?
 - [ ] Does the new code have enough tests? (*every* PR should have tests or justification otherwise. Bug-fix PRs especially)
 - [ ] Does the new code have enough documentation that answers "how do I use it?" and "what does it do?"? (both source documentation and [higher level](https://github.com/storj/docs), diagrams?)
 - [ ] Does any documentation need updating?