@wthorp 2019-05-16
Cam was working on optimizing this query... not sure if this PR is a response to his work or not.