@cam-a 2019-05-16
@wthorp @egonelbre I didn't end up making the changes to that query